// lib/generated/userEnums.ts
// 🚨 AUTO-GENERATED FILE. DO NOT EDIT MANUALLY!
// Generated from Prisma schema.


export enum UserRole {
  ADMIN = "ADMIN",
  THERAPIST = "THERAPIST",
  RECEPTIONIST = "RECEPTIONIST",
  CONTENT_MANAGER = "CONTENT_MANAGER",
}


export enum UserStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  SUSPENDED = "SUSPENDED",
}


export enum Gender {
  MALE = "MALE",
  FEMALE = "FEMALE",
  OTHER = "OTHER",
}